Remove useWindowDimensions hook
